PLATOPS-1542 - Migrated to new build
