+++ build.sbt
@@ -1,0 +1,14 @@
+import play.sbt.PlayImport.PlayKeys.devSettings
+import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin.publishingSettings
+
+val appName = "service-deployments"
+
+lazy val microservice = Project(appName, file("."))
+  .enablePlugins(
+    Seq(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin, SbtArtifactory): _*)
+  .settings(majorVersion := 1)
+  .settings(publishingSettings: _*)
+  .settings(devSettings := Seq("play.server.netty.maxInitialLineLength" -> "65536"))
+  .settings(PlayKeys.playDefaultPort := 8458)
+  .settings(libraryDependencies ++= AppDependencies.compile ++ AppDependencies.test)
+  .settings(resolvers += Resolver.jcenterRepo)
