--- project/MicroService.scala
@@ -1,56 +1,0 @@
-import play.sbt.PlayImport.PlayKeys._
-import sbt.Keys._
-import sbt.Tests.{Group, SubProcess}
-import sbt._
-import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin._
-import play.sbt.routes.RoutesKeys.{routesGenerator, _}
-
-trait MicroService {
-
-  import uk.gov.hmrc._
-  import DefaultBuildSettings._
-  import uk.gov.hmrc.{SbtBuildInfo, ShellPrompt}
-
-  import TestPhases._
-
-  val appName: String
-
-  lazy val appDependencies: Seq[ModuleID] = ???
-  lazy val plugins: Seq[Plugins]          = Seq()
-  lazy val playSettings: Seq[Setting[_]]  = Seq.empty
-
-  lazy val microservice = Project(appName, file("."))
-    .enablePlugins(Seq(play.sbt.PlayScala) ++ plugins: _*)
-    .settings(playSettings: _*)
-    .settings(devSettings := Seq("play.server.netty.maxInitialLineLength" -> "65536"))
-    .settings(scalaSettings: _*)
-    .settings(publishingSettings: _*)
-    .settings(defaultSettings(): _*)
-    .settings(
-      targetJvm := "jvm-1.8",
-      scalaVersion := "2.11.11",
-      libraryDependencies ++= appDependencies,
-      parallelExecution in Test := false,
-      fork in Test := false,
-      routesGenerator := InjectedRoutesGenerator,
-      evictionWarningOptions in update := EvictionWarningOptions.default.withWarnScalaVersionEviction(false)
-    )
-    .configs(IntegrationTest)
-    .settings(inConfig(IntegrationTest)(Defaults.itSettings): _*)
-    .settings(
-      Keys.fork in IntegrationTest := false,
-      unmanagedSourceDirectories in IntegrationTest <<= (baseDirectory in IntegrationTest)(base => Seq(base / "it")),
-      addTestReportOption(IntegrationTest, "int-test-reports"),
-      testGrouping in IntegrationTest := oneForkedJvmPerTest((definedTests in IntegrationTest).value),
-      parallelExecution in IntegrationTest := false
-    )
-    .settings(resolvers ++= Seq(Resolver.bintrayRepo("hmrc", "releases"), Resolver.jcenterRepo))
-}
-
-private object TestPhases {
-
-  def oneForkedJvmPerTest(tests: Seq[TestDefinition]) =
-    tests map { test =>
-      new Group(test.name, Seq(test), SubProcess(ForkOptions(runJVMOptions = Seq("-Dtest.name=" + test.name))))
-    }
-}
